--- a/.gitignore
+++ b/.gitignore
@@ -9,3 +9,4 @@ storybook-static/
 *.tsbuildinfo
 .vscode/
 *.tgz
+.swc/