--- a/.storybook/main.js
+++ b/.storybook/main.js
@@ -1,4 +1,8 @@
 import { dirname, join } from "path";
+import { mergeConfig } from "vite";
+import ViteYaml from '@modyfi/vite-plugin-yaml';
+import graphqlLoader from 'vite-plugin-graphql-loader';
+
 const path = require("path");
 
 module.exports = {
@@ -7,8 +11,11 @@ module.exports = {
     getAbsolutePath("@storybook/addon-docs"),
     getAbsolutePath("@storybook/addon-links"),
     "storybook-react-intl",
-    getAbsolutePath("@storybook/addon-webpack5-compiler-babel")
   ],
+
+  core: {
+    builder: '@storybook/builder-vite',
+  },
 
   stories: [
     "../packages/*/src/**/*.story.@(js|jsx|ts|tsx)"
@@ -16,58 +23,41 @@ module.exports = {
 
   staticDirs: ['../public'],
 
-  webpackFinal: async (config, { configType }) => {
-    // This method is for altering Storybook's webpack configuration.
-    // Add support for importing image files
-    config.module.rules.push({
-      test: /\.(png|jpg|gif|svg)$/,
-      use: [
-        {
-          loader: 'file-loader',
-        },
+  async viteFinal(config, { configType }) {
+    // This method is for altering Storybook's Vite configuration.
+    return mergeConfig(config, {
+      plugins: [
+        // Add support for importing YAML files
+        ViteYaml(),
+        // Add support for importing GraphQL files
+        graphqlLoader(),
       ],
-      include: path.resolve(__dirname, './packages/transitive-overlay/src/images'),
+      
+      assetsInclude: [
+        // Add support for importing image files
+        '**/*.png',
+        '**/*.jpg',
+        '**/*.gif',
+        '**/*.svg'
+      ],
+
+      // Add fallback for querystring
+      define: {
+        global: 'globalThis',
+      },
+
+      resolve: {
+        alias: {
+          // Configure module resolution for workspace packages
+          '@opentripplanner': path.resolve(__dirname, '../packages'),
+          querystring: 'querystring-es3'
+        }
+      },
     });
-    // Add support for importing YAML files.
-    config.module.rules.push({
-      test: /\.(yml|yaml)$/,
-      loader: "yaml-loader"
-    });
-
-    config.module.rules.push({
-      test: /\.graphql$/,
-      exclude: /node_modules/,
-      loader: 'graphql-tag/loader',
-    });
-
-    config.module.rules.push({
-      test: /uFuzzy/,
-      loader: 'babel-loader',
-      options: {
-        presets: [
-          ['@babel/preset-env', { targets: 'defaults' }]
-        ]
-      }
-    })
-
-    // Add fallback for querystring
-    config.resolve.fallback = {
-      ...config.resolve.fallback,
-      querystring: require.resolve('querystring-es3')
-    };
-
-    // Configure module resolution for workspace packages
-    config.resolve.alias = {
-      ...config.resolve.alias,
-      '@opentripplanner': path.resolve(__dirname, '../packages')
-    };
-
-    // Return the altered config
-    return config;
   },
 
   framework: {
-    name: getAbsolutePath("@storybook/react-webpack5"),
+    name: getAbsolutePath("@storybook/react-vite"),
     options: {}
   },