--- a/.swcrc
+++ b/.swcrc
@@ -1,0 +1,49 @@
+{
+  "$schema": "https://swc.rs/schema.json",
+  "jsc": {
+    "parser": {
+      "syntax": "typescript",
+      "tsx": true,
+      "dynamicImport": true,
+      "privateMethod": true,
+      "functionBind": false,
+      "exportDefaultFrom": false,
+      "exportNamespaceFrom": false,
+      "decorators": false,
+      "decoratorsBeforeExport": false,
+      "topLevelAwait": false,
+      "importMeta": false
+    },
+    "transform": {
+      "react": {
+        "runtime": "automatic",
+        "pragma": "React.createElement",
+        "pragmaFrag": "React.Fragment",
+        "throwIfNamespace": true,
+        "development": false,
+        "useBuiltins": false
+      },
+      "useDefineForClassFields": true,
+      "legacyDecorator": false,
+      "decoratorMetadata": false
+    },
+    "target": "es2018",
+    "loose": true,
+    "externalHelpers": false,
+    "keepClassNames": false,
+    "experimental": {
+      "plugins": [
+        ["@swc/plugin-styled-components", {}]
+      ]
+    }
+  },
+  "module": {
+    "type": "commonjs",
+    "strict": false,
+    "strictMode": true,
+    "lazy": false,
+    "noInterop": false
+  },
+  "minify": false,
+  "isModule": true
+}