--- a/jest.config.js
+++ b/jest.config.js
@@ -1,7 +1,10 @@
 module.exports = {
   cacheDirectory: ".jest-cache",
   coverageDirectory: ".jest-coverage",
-  coveragePathIgnorePatterns: ["<rootDir>/packages/(?:.+?)/lib/"],
+  coveragePathIgnorePatterns: [
+    "<rootDir>/packages/(?:.+?)/lib/",
+    "<rootDir>/packages/(?:.+?)/esm/"
+  ],
   coverageReporters: ["html", "text"],
   coverageThreshold: {
     global: {
@@ -31,6 +34,7 @@ module.exports = {
   transform: {
     "\\.[jt]sx?$": "@swc/jest",
     "\\.ya?ml$": "yaml-jest",
+    "\\.graphql$": "jest-transform-graphql",
     "node_modules/chroma-js/.+\\.js$": "<rootDir>/transform-esm-modules.js"
   },
   transformIgnorePatterns: ["/node_modules/(?!(chroma-js)/)"]