--- a/packages/base-map/__mocks__/AllVehicles.tsx
+++ b/packages/base-map/__mocks__/AllVehicles.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Marker } from "react-map-gl";
 import { LeafletStyleMarker, Popup } from "../src/styled";
 
-const vehicleData = require("./vehicle-data/all-trimet.json");
+import vehicleData from "./vehicle-data/all-trimet.json";
 
 type Props = {
   id?: string;