--- a/packages/building-blocks/src/stories/dropdown.story.tsx
+++ b/packages/building-blocks/src/stories/dropdown.story.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ComponentMeta } from "@storybook/react";
+import { ComponentMeta } from "@storybook/react-webpack5";
 import styled from "styled-components";
 import Dropdown from "../dropdown";
 import blue from "../colors/blue";