--- a/packages/building-blocks/src/stories/index.story.tsx
+++ b/packages/building-blocks/src/stories/index.story.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from "react";
-import { Meta } from "@storybook/react";
+import { Meta } from "@storybook/react-webpack5";
 import ColorPalette from "./ColorPalette";
 
 import grey from "../colors/grey";