--- a/packages/from-to-location-picker/src/index.story.tsx
+++ b/packages/from-to-location-picker/src/index.story.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { IntlProvider } from "react-intl";
-import { action } from "@storybook/addon-actions";
+import { action } from "storybook/actions";
 
 import FromToLocationPicker from ".";