--- a/packages/itinerary-body/src/defaults/route-description.tsx
+++ b/packages/itinerary-body/src/defaults/route-description.tsx
@@ -1,10 +1,12 @@
-import { getLegRouteShortName } from "@opentripplanner/core-utils/lib/itinerary";
+import coreUtils from "@opentripplanner/core-utils";
 import React, { ReactElement } from "react";
 
 import * as S from "../styled";
 import { RouteDescriptionProps } from "../types";
 
 import RouteLongName from "./route-long-name";
+
+const { getLegRouteShortName } = coreUtils.itinerary;
 
 export default function RouteDescription({
   leg