--- a/packages/itinerary-body/src/stories/OtpUiItineraryBody.story.tsx
+++ b/packages/itinerary-body/src/stories/OtpUiItineraryBody.story.tsx
@@ -2,7 +2,7 @@ import React, { ReactElement } from "react";
 import { Bomb } from "@styled-icons/fa-solid/Bomb";
 import { Bolt } from "@styled-icons/fa-solid/Bolt";
 import styled from "styled-components";
-import { Meta } from "@storybook/react";
+import { Meta } from "@storybook/react-webpack5";
 import RouteDescriptionFooterWithWaitTimes from "./footer-with-wait-times";
 
 import ItineraryBody from "..";