--- a/packages/location-field/src/stories/Mobile.story.tsx
+++ b/packages/location-field/src/stories/Mobile.story.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ComponentMeta } from "@storybook/react";
+import { ComponentMeta } from "@storybook/react-webpack5";
 import styled from "styled-components";
 import {
   Building,