--- a/packages/printable-itinerary/src/tnc-leg.tsx
+++ b/packages/printable-itinerary/src/tnc-leg.tsx
@@ -1,6 +1,5 @@
-import { getCompanyForNetwork } from "@opentripplanner/core-utils/lib/itinerary";
-import { Defaults } from "@opentripplanner/itinerary-body";
-import { parseOTP2Minute } from "@opentripplanner/itinerary-body/lib/util";
+import coreUtils from "@opentripplanner/core-utils";
+import { Defaults, parseOTP2Minute } from "@opentripplanner/itinerary-body";
 import { GradationMap, Leg, LegIconComponent } from "@opentripplanner/types";
 import React, { ReactElement } from "react";
 import { FormattedMessage } from "react-intl";
@@ -8,6 +7,8 @@ import { FormattedMessage } from "react-intl";
 import AccessibilityAnnotation from "./accessibility-annotation";
 import * as S from "./styled";
 import { defaultMessages, strongText } from "./util";
+
+const { getCompanyForNetwork } = coreUtils.itinerary;
 
 interface Props {
   accessibilityScoreGradationMap?: GradationMap;