--- a/packages/zoom-based-markers/src/ZoomBasedMarker.story.tsx
+++ b/packages/zoom-based-markers/src/ZoomBasedMarker.story.tsx
@@ -14,7 +14,7 @@ import {
   SymbolComponentBaseProps,
   ZoomBasedSymbol
 } from "@opentripplanner/types";
-import { action } from "@storybook/addon-actions";
+import { action } from "storybook/actions";
 import React, { useState } from "react";
 
 import ZoomBasedMarkers from ".";