… .storybook/base-map-wrapper.jsx
… .storybook/manager.js
… babel.config.json
… package.json
… packages/icons/src/standard/standard-flex.jsx
… pnpm-lock.yaml