This PR is an experiment in switching some outdated build system tools over to modern replacements that are much faster. 

Some time comparisons for building:
| Command | Before Time | After Time |
|--------|--------|--------|
| `pnpm build:cjs` | 4.43s | 3.18s |
| `pnpm build:esm` | 5.01s | 2.86s |
| `pnpm build-storybook` | 24.78s | 20.61s | 
| `pnpm dev` time to preview | 17s | 3.33s |
| `pnpm unit` (no cache) | 5.118s | 1.9s |

A little disappointed the speedup in building wasn't better, but I think that's because the build process itself is quite fast and most of the time is spend in lerna spinning up all the processes. The biggest improvement is in the storybook dev, because vite uses modern web browsers' support for esm natively it doesn't have to compile stuff before send it to the browser, so it can be much faster.